+++ src/main/java/vontus/magicbottle/util/Utils.java
@@ -1,4 +1,4 @@
-package vontus.magicbottle;
+package vontus.magicbottle.util;
 
 import java.text.NumberFormat;
 import java.util.Locale;
